functions plotData, prepareForPlotting moved to './plottingUtils.js'
