+++ utils/JSONGrapherMainScript.js
@@ -1,11 +1,7 @@
       import { jsonifyData, findFileType, createCSV, getFileName, readFileAsText } from './fileUtils.js'; 
       import {initializeUniversalSchemas, getSchemaType, mergeFigDictWithTemplate, getSchemaLocation, validateData} from './schemaUtils.js'
-      import {getUnitFromLabel, removeUnitFromLabel, replaceSuperscripts} from './unitUtils.js'
-      import {convertUnits} from './figDictUtils.js'
-      import {executeImplicitDataSeriesOperations} from './json_equationer/implicitUtils.js'
-      import { parsePlotStyle, applyPlotStyleToPlotlyDict } from './styleUtils.js';
-      import { cleanJsonFigDict } from './figDictUtils.js'; 
       import { loadJsonFromUrl, isValidUrl, parseUrl } from './linkUtils.js'; 
+      import {plotData, prepareForPlotting} from './plottingUtils.js'
 
       function copyJson(obj) { //for debugging.
         return JSON.parse(JSON.stringify(obj));
@@ -55,39 +51,7 @@
         Plotly.purge("plotlyDiv");
       }
 
-      // A function that visualizes the data with plotly
-      async function plot_with_plotly(figDict) {
-        let plotStyle = { layout_style: "", trace_styles_collection: "" };
-          if (JSON.stringify(plotStyle) === JSON.stringify({ layout_style: "", trace_styles_collection: "" })) {
-              plotStyle = figDict.plot_style ?? { layout_style: "", trace_styles_collection: "" };
-          }
 
-          let copyForPlotly = JSON.parse(JSON.stringify(figDict)); // Plotly mutates the input, and we also do when applying plot style.      
-          //offset2D and arrange2dTo3d must be executed after making the copy, if requested. The other implicit functions have already been called.
-          copyForPlotly = await executeImplicitDataSeriesOperations(copyForPlotly, false, false, false, true, true);
-          // Parse the plot style
-          plotStyle = parsePlotStyle(plotStyle);
-          // Apply the plot style
-          copyForPlotly = applyPlotStyleToPlotlyDict(copyForPlotly, plotStyle);
-
-          // Clean out the fields to make a Plotly figDict
-          copyForPlotly = cleanJsonFigDict(copyForPlotly, ['simulate', 'custom_units_chevrons', 'equation', 'trace_style', '3d_axes', 'bubble', 'superscripts', 'nested_comments', 'extraInformation']);          
-
-          // Replace superscripts in some fields for Plotly
-          if (copyForPlotly.layout.scene) {
-            copyForPlotly.layout.scene.yaxis.title.text = replaceSuperscripts(copyForPlotly.layout.scene.yaxis.title?.text ?? "");
-            copyForPlotly.layout.scene.xaxis.title.text = replaceSuperscripts(copyForPlotly.layout.scene.xaxis.title?.text ?? "");
-          } else {
-            copyForPlotly.layout.yaxis.title.text = replaceSuperscripts(copyForPlotly.layout.yaxis.title?.text ?? "");
-            copyForPlotly.layout.xaxis.title.text = replaceSuperscripts(copyForPlotly.layout.xaxis.title?.text ?? "");
-          }
-          // Ensure Plotly is available before calling newPlot
-          if (typeof Plotly !== "undefined") {
-              await Plotly.newPlot("plotlyDiv", copyForPlotly.data, copyForPlotly.layout);
-          } else {
-              console.error("Plotly is not loaded.");
-          }
-      }
 
       function createCopyURLButton(jsonURL) {
           // Generate the URL
@@ -124,7 +88,6 @@
       }
 
       // A function that will create a download button for the csv file
-      //As of 6/15/2025, this file only contains data from the most recently uploaded dataset.
       function createDownloadCSVButton(csv, filename) {
         // Creating a download link for the csv file
         const downloadLink = createDownloadCSVLink(csv.csv, filename);
@@ -137,7 +100,6 @@
       }
 
       //A function that will create a download button for the JSON file
-      //As of 6/15/2025, this file only contains data from the most recently uploaded dataset.
       // For the first argument, this takes in a json object (a "javascript object", not a string)
       // For the second argument, the filename field is a string.
       function createDownloadJSONButton(json, filename) {
@@ -176,7 +138,7 @@
         // insert a download button with downloadLink after downloadJSON
         // Clear the container
         buttonsContainer.innerHTML =
-          "&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp; Download Last Data Set As:"; //&nbsp; is HTML code to add a space.
+          "&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp; Download Graph Record As:"; //&nbsp; is HTML code to add a space.
         // Add download JSON button
         buttonsContainer.appendChild(downloadJSONButton);
         // Add download CSV button
@@ -313,7 +275,10 @@
         if (!_jsonified) return;
 
         globalData = await plotData(globalData, _jsonified, recentFileName, messagesToUserDiv, errorDiv); // STEP 4–7
-
+          // STEP 6: Provide file with converted units for download as JSON and CSV by buttons
+          //should  make an if statement here to give newestFigDict with filename if only one record has been uploaded
+          // and to otherwise give the full data with name like "mergedGraphRecord.json" for the filename.
+        if (globalData){appendDownloadButtons(globalData, "mergedGraphRecord.json");}
         errorDiv.innerText = errorDiv.innerText.replace(loadingMessage, "");
       }
 
@@ -374,113 +339,5 @@
         return { jsonified, fileType };
       }
 
-
-      // If the data is valid against the schema, then we can proceed to the next step
-      // if necessary create download button with json
-      async function plotData(globalData, _jsonified, recentFileName, messagesToUserDiv, errorDiv) {
-        // STEP 4 and STEP 5 is done in the prepareForPlotting function
-        const { mergedFigDict, fileName, newestFigDict } = await prepareForPlotting(globalData, _jsonified, recentFileName, errorDiv); // recentFileName is a global variable.
-        if (mergedFigDict) {
-          // STEP 6: Provide file with converted units for download as JSON and CSV by buttons
-          //should  make an if statement here to give newestFigDict with filename if only one record has been uploaded
-          // and to otherwise give the full data with name like "mergedGraphRecord.json" for the filename.
-          appendDownloadButtons(newestFigDict, fileName);
-
-          // STEP 7: Then create a plotly JSON, clean it, and render it on the browser
-          plot_with_plotly(mergedFigDict);
-          //Replace existing "Data Plotted" message if it is already there, to avoid duplicating it.
-          if (!messagesToUserDiv.innerText) { //Currently, we assume the below message is present or not present. If we later put additional messagesToUser, we may need to add more logic.
-            const dataPlottedMessage = "\u2003\u2003\u2003\u2003\u2003\u2003 Data plotted! Add more data or click 'Clear Data' to start a new graph! \u2003\u2003\u2003\u2003\u2003\u2003"
-            messagesToUserDiv.innerText += dataPlottedMessage;
-          }
-          return mergedFigDict //This returns the mergedFigDict to use as globalData.
-        } else {
-          console.log("Plotting skipped: incompatible data or merge failure.");
-          return;
-        }
-      }
-
-      // This a function that plots the data on the graph
-      // the input, jsonified, is the new figDict. globalData is the 'global' figDict.
-      async function prepareForPlotting(globalData, jsonified, recentFileName, errorDiv) {
-        try {
-          // Checks if the Jsonified is the first file uploaded
-          if (!globalData) {
-            let _jsonified = JSON.parse(JSON.stringify(jsonified)); // make a local copy
-            globalData = copyJson(_jsonified); // populate global figDict since this is the first record received.            
-
-            // Get the unit from the label
-            const xUnit = getUnitFromLabel(_jsonified.layout.xaxis.title.text);
-            const yUnit = getUnitFromLabel(_jsonified.layout.yaxis.title.text);
-
-            // Adding the extracted units to _jsonified
-            _jsonified.unit = {
-              x: xUnit,
-              y: yUnit,
-            };
-
-            // STEP 4: Check if the object has a dataSet that has a simulate key in it, and runs the simulate function based on the value provided in the key model
-            _jsonified = await executeImplicitDataSeriesOperations(_jsonified); // _jsonified is a figDict.
-
-            // No STEP 5 for first record; it defines the units of globalData.
-            globalData = _jsonified;
-
-            // Return the objects that have been prepared for plotting and downloading.
-            return {
-              mergedFigDict: globalData,
-              newestFigDict: _jsonified,
-              fileName: recentFileName,
-            };
-          } else {
-            let fieldsMatch = true;
-
-            if (globalData.datatype !== jsonified.datatype) {
-              fieldsMatch = false;
-              errorDiv.innerText += "The added record's datatype is different. Stopping merging. The two values are: " +
-                String(globalData.datatype) + " " + String(jsonified.datatype) + "\n";
-            }
-
-            if (removeUnitFromLabel(globalData.layout.xaxis.title.text) !==
-                removeUnitFromLabel(jsonified.layout.xaxis.title.text)) {
-              fieldsMatch = false;
-              errorDiv.innerText += "The added record's xaxis label text is different. Stopping merging. The two values are: " +
-                removeUnitFromLabel(globalData.layout.xaxis.title.text) + " " +
-                removeUnitFromLabel(jsonified.layout.xaxis.title.text) + "\n";
-            }
-
-            if (removeUnitFromLabel(globalData.layout.yaxis.title.text) !==
-                removeUnitFromLabel(jsonified.layout.yaxis.title.text)) {
-              fieldsMatch = false;
-              errorDiv.innerText += "The added record's yaxis label text is different. Stopping merging. The two values are: " +
-                removeUnitFromLabel(globalData.layout.yaxis.title.text) + " " +
-                removeUnitFromLabel(jsonified.layout.yaxis.title.text) + "\n";
-            }
-
-            if (fieldsMatch) {
-              let _jsonified = JSON.parse(JSON.stringify(jsonified));
-              _jsonified = await executeImplicitDataSeriesOperations(_jsonified);
-              _jsonified = await convertUnits(_jsonified, globalData);
-
-              globalData.data = [
-                ...globalData.data,
-                ..._jsonified.data,
-              ];
-
-              return {
-                mergedFigDict: globalData,
-                newestFigDict: _jsonified,
-                fileName: recentFileName,
-              };
-            } else {
-              errorDiv.innerText += `Added data not plotted. The records were not compatible for merging. You may continue trying to add data sets, or may click "Clear Data" to start a new graph. These error messages will be automatically cleared after 10 seconds. \n`;
-              setTimeout(() => { errorDiv.innerText = ''; }, 10000);
-              return null;
-            }
-          }
-        } catch (err) {
-          console.log("Error from plotData: ", err);
-          throw err;
-        }
-      }
 
 window.clearData = clearData;
